fix lengths for va_start and CF2 large arity calls

+++ include/otfcc/primitives.h
@@ -11,7 +11,7 @@ typedef uint16_t glyphclass_t;
 typedef uint16_t tableid_t;
 typedef uint16_t shapeid_t;
 typedef uint16_t cffsid_t;
-typedef uint16_t arity_t;
+typedef uint32_t arity_t; // CFF2 may support large arity
 
 typedef double pos_t;
 typedef double length_t;
